extract useOmniYieldLoopHeadline
