+++ features/omni-kit/hooks/index.ts
@@ -2,3 +2,4 @@ export * from './useOmniProtocolData'
 export * from './useOmniTransition'
 export * from './useOmniTxHandler'
 export * from './useOmniSlippage'
+export * from './useOmniYieldLoopHeadline'
